+++ react/helpers/index.tsx
@@ -1,8 +1,14 @@
-export async function fetchEmail() {
-  const { Email: email = '' } = await fetch(
-    '/no-cache/profileSystem/getProfile'
-  ).then(res => res.json())
-  return email
+export async function fetchEmail(): Promise<string> {
+  try {
+    const { Email: email } = await (await fetch(
+      '/no-cache/profileSystem/getProfile'
+    )).json()
+    return email
+  } catch (err) {
+    console.log(`Couldn't fetch logged in use email`)
+    console.log(err)
+  }
+  return ''
 }
 
 export function getSiteType() {